+++ src/guidellm/data/deserializers/memory.py
@@ -1,10 +1,8 @@
 from __future__ import annotations
 
 from collections.abc import Callable
-from typing import Any, Literal
 
 from datasets import Dataset
-from pydantic import Field
 from transformers import PreTrainedTokenizerBase
 
 from guidellm.data.deserializers.deserializer import (
@@ -12,7 +10,11 @@ from guidellm.data.deserializers.deserializer import (
     DatasetDeserializer,
     DatasetDeserializerFactory,
 )
-from guidellm.data.schemas import DataArgs
+from guidellm.schemas.data.deserializers import (
+    InMemoryDictDataArgs,
+    InMemoryDictListDataArgs,
+    InMemoryItemListDataArgs,
+)
 
 __all__ = [
     "InMemoryDictDataArgs",
@@ -22,20 +24,6 @@ __all__ = [
     "InMemoryItemListDataArgs",
     "InMemoryItemListDatasetDeserializer",
 ]
-
-
-@DataArgs.register("in_memory_dict")
-class InMemoryDictDataArgs(DataArgs):
-    """Model for in-memory data deserializer arguments."""
-
-    kind: Literal["in_memory_dict"] = Field(  # type: ignore[assignment]
-        default="in_memory_dict",
-        description="Type identifier for the in-memory data deserializer.",
-    )
-    data: dict[str, list] = Field(
-        description="In-memory data input for the dataset deserializer.",
-        examples=[{"column1": [1, 2, 3], "column2": [4, 5, 6]}],
-    )
 
 
 @DatasetDeserializerFactory.register("in_memory_dict")
@@ -62,18 +50,6 @@ class InMemoryDictDatasetDeserializer(DatasetDeserializer):
             )
 
         return Dataset.from_dict(data, **config.load_kwargs)
-
-
-@DataArgs.register("in_memory_dict_list")
-class InMemoryDictListDataArgs(DataArgs):
-    kind: Literal["in_memory_dict_list"] = Field(  # type: ignore[assignment]
-        default="in_memory_dict_list",
-        description="Type identifier for the in-memory data deserializer.",
-    )
-    data: list[dict[str, Any]] = Field(
-        description="In-memory list of dicts input for the dataset deserializer.",
-        examples=[{"column1": 1, "column2": 2}, {"column1": 3, "column2": 4}],
-    )
 
 
 @DatasetDeserializerFactory.register("in_memory_dict_list")
@@ -107,22 +83,6 @@ class InMemoryDictListDatasetDeserializer(DatasetDeserializer):
                 result_dict[key].append(value)
 
         return Dataset.from_dict(result_dict, **config.load_kwargs)
-
-
-@DataArgs.register("in_memory_item_list")
-class InMemoryItemListDataArgs(DataArgs):
-    kind: Literal["in_memory_item_list"] = Field(  # type: ignore[assignment]
-        default="in_memory_item_list",
-        description="Type identifier for the in-memory data deserializer.",
-    )
-    data: list[str | int | float | bool | None] = Field(
-        description="In-memory list of primitive items for the dataset deserializer.",
-        examples=[[1, 2, 3, 4, 5]],
-    )
-    column_name: str = Field(
-        default="data",
-        description="Column name to use when creating the dataset.",
-    )
 
 
 @DatasetDeserializerFactory.register("in_memory_item_list")
