+++ src/guidellm/data/preprocessors/encoders.py
@@ -1,18 +1,16 @@
 from __future__ import annotations
 
-from typing import Any, Literal
-
-from pydantic import Field
+from typing import Any
 
 from guidellm.data.preprocessors.preprocessor import (
     DatasetPreprocessor,
     PreprocessorRegistry,
 )
-from guidellm.data.schemas import DataPreprocessorArgs
+from guidellm.schemas.data.preprocessors import MediaEncoderArgs
 from guidellm.utils import audio as guidellm_audio
 from guidellm.utils import vision as guidellm_vision
 
-__all__ = ["MediaEncoder"]
+__all__ = ["MediaEncoder", "MediaEncoderArgs"]
 
 
 def _is_encoded_image(item: Any) -> bool:
@@ -21,31 +19,6 @@ def _is_encoded_image(item: Any) -> bool:
 
 def _is_encoded_video(item: Any) -> bool:
     return isinstance(item, dict) and "video" in item and "type" in item
-
-
-@DataPreprocessorArgs.register("encode_media")
-class MediaEncoderArgs(DataPreprocessorArgs):
-    """Model for media encoder preprocessor arguments."""
-
-    kind: Literal["encode_media"] = Field(
-        default="encode_media",
-        description="Type identifier for the media encoder preprocessor.",
-    )
-    audio_kwargs: dict[str, Any] = Field(
-        default_factory=dict,
-        description="Keyword arguments for audio encoding.",
-        examples=[{"format": "mp3"}],
-    )
-    image_kwargs: dict[str, Any] = Field(
-        default_factory=dict,
-        description="Keyword arguments for image encoding.",
-        examples=[{"format": "jpg"}],
-    )
-    video_kwargs: dict[str, Any] = Field(
-        default_factory=dict,
-        description="Keyword arguments for video encoding.",
-        examples=[{"format": "mp4"}],
-    )
 
 
 @PreprocessorRegistry.register("encode_media")
