+++ src/guidellm/schemas/data/finalizers/__init__.py
@@ -1,0 +1,7 @@
+from __future__ import annotations
+
+from guidellm.schemas.data.finalizers.generative import (
+    GenerativeRequestFinalizerArgs,
+)
+
+__all__ = ["GenerativeRequestFinalizerArgs"]
